Nicolas Regnault: - add support of long int in SingleIntegerOption
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1014 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Options/OptionManager.cc
@@ -471,7 +471,7 @@ double* OptionManager::GetDoubles(const char *optionName, int & length)
 }
 
 //accessor routine for Integer value
-int OptionManager::GetInteger(const char *optionName)
+long OptionManager::GetInteger(const char *optionName)
   {
   AbstractOption* OptionPointer = (*this)[optionName];
   if (OptionPointer!=0)
